Update to v1.1.0
## [1.1.0] - 2017-02-20
### Added
* Cannonballs are now vector based instead of array based
* Path is now vector based
* Random colors are now seeded based on time. Because reasons...

### Updated
* Many variable names to be more in line with [Google's style
guide](https://google.github.io/styleguide/cppguide.html)
* Various other changes to make it more in line with the style guide
* Various other improvements to code function and variable types
* Number of cannonballs is now limited by computer memory
* Window values are now held more globally so I don't have to keep
passing it everything that needs it.
This likely isn't the best method to achieve it, but it prevents an
issue with switching to vector based cannonballs.\
* Max number of past points is now a config value

### Notes
* There are still quite a few non-standard elements left from before I
started using Google's style guide consistantly.
I changed the worst offending members, but most of the rest of it will
be left.